Added ChatMessage component and fix for user's message
