+++ styles/globals.css
@@ -48,6 +48,13 @@
   --feature-card-twitter: #1792e8;
 }
 
+h1,
+h2,
+p,
+.chatbox {
+  font-family: "Montserrat" !important;
+}
+
 @media (prefers-color-scheme: dark) {
   :root {
     --foreground-rgb: 255, 255, 255;
